requirements.txt
pyyaml==6.0.0
python-telegram-bot==13.11.0
paramiko==2.10.3
